+++ library/src/commonTest/kotlin/io/github/ezer_mackenzie/proximitytransfer/core/transfer/config/TransferLimitsTest.kt
@@ -1,0 +1,16 @@
+package io.github.ezer_mackenzie.proximitytransfer.core.transfer.config
+
+import kotlin.test.Test
+import kotlin.test.assertFailsWith
+
+class TransferLimitsTest {
+    @Test
+    fun rejectsInvalidLimits() {
+        assertFailsWith<IllegalArgumentException> {
+            TransferLimits(maxPayloadBytes = -1)
+        }
+        assertFailsWith<IllegalArgumentException> {
+            TransferLimits(maxChunkCount = 0)
+        }
+    }
+}